requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
python-dotenv
google-generativeai
markdown
PyPDF2
pdfminer.six
